more codeblock rules
minor refactor

use sandpack

add rust support

custom entry point

file tree

style the editor for single and multiple snippets
